chore(controllers): feedback to follow conventions in naming - rename controllers file and class name to pascal namings - change the entry point for route to index - follow documentations for response messages [Finishes #163310609]
