+++ server/database/dummydbParties.js
@@ -2,19 +2,19 @@ const parties = [
   {
     id: 1,
     name: 'APC',
-    hqaddress: '123 williams Akins, Lagos, Nigeria',
+    hqAddress: '123 williams Akins, Lagos, Nigeria',
     logoUrl: 'Image Url 1 goes here'
   },
   {
     id: 2,
     name: 'PDP',
-    hqaddress: '213 williams Akins, Lagos, Nigeria',
+    hqAddress: '213 williams Akins, Lagos, Nigeria',
     logoUrl: 'Image Url 2 goes here'
   },
   {
     id: 3,
     name: 'APGA',
-    hqaddress: '321 williams Akins, Lagos, Nigeria',
+    hqAddress: '321 williams Akins, Lagos, Nigeria',
     logoUrl: 'Image Url 3 goes here'
   }
 ];
